Use dashboard component for logged in & public sharing views
