+++ dev/scripts/app.js
@@ -13,7 +13,6 @@ import Dashboard from './components/dashboard.js'
 import SingleApplication from './components/jobApplication';
 import Home from './components/home';
 import NewApplication from './components/newApplication.js'
-import SharedDashboard from './components/sharedDashboard';
 import SharedSingleApplication from './components/sharedJobApplication';
 
 // firebase config
@@ -118,8 +117,9 @@ class App extends React.Component {
                 <Route exact path='/application/:application_id' render={(routeProps) => {
                     return <SingleApplication {...routeProps} userId={this.state.userId} />
                 }} />
+                {/* Public Sharing Routes */}
                 <Route exact path='/shared/:userId/:shareKey' render={(routeProps) => {
-                  return <SharedDashboard {...routeProps} isLoggedIn={this.state.isLoggedIn} />
+                  return <Dashboard {...routeProps} isLoggedIn={this.state.isLoggedIn} isSharedView='true' />
                 }} />
                 <Route exact path='/shared/:userId/:shareKey/:application_id' component={SharedSingleApplication} />
                 {/* If no paths match, display an error message */}
@@ -133,8 +133,9 @@ class App extends React.Component {
               // Routes if the user is logged out
             : <Switch>
                 <Route exact path='/' component={Home} />
+                {/* Public Sharing Routes */}
                 <Route exact path='/shared/:userId/:shareKey' render={(routeProps) => {
-                  return <SharedDashboard {...routeProps} isLoggedIn={this.state.isLoggedIn} isSharedView='true' />
+                  return <Dashboard {...routeProps} isLoggedIn={this.state.isLoggedIn} isSharedView='true' />
                 }} />
                 <Route exact path='/shared/:userId/:shareKey/:application_id' component={SharedSingleApplication} />
                 {/* If no paths match, display an error message */}
